third_party/blink/web_tests/external/wpt/fetch/corb/response_block.tentative.https-expected.txt

This is a testharness.js-based test.
[FAIL] ORB: Expect error response from fetch().
  assert_equals: expected "ORB-style network error" but got "Resource loaded (expected for CORB)"
Harness: the test ran to completion.
